+++ vimiv/completion/completer.py
@@ -22,42 +22,42 @@ class Completer(QObject):
     Attributes:
         _cmd: CommandLine object.
         _completion: CompletionWidget object.
-        _proxy_model: The completion filter used.
     """
 
     @api.objreg.register
     def __init__(self, commandline, completion):
         super().__init__()
 
-        self._proxy_model = None
         self._cmd = commandline
         self._completion = completion
 
-        self._completion.activated.connect(self._on_completion)
+        self._completion.activated.connect(self._complete)
         api.modes.COMMAND.first_entered.connect(self._init_models)
-        api.modes.COMMAND.entered.connect(self._on_entered)
-        api.modes.COMMAND.left.connect(self._on_left)
         self._cmd.textEdited.connect(self._on_text_changed)
         self._cmd.editingFinished.connect(self._on_editing_finished)
+
+    @property
+    def proxy_model(self) -> api.completion.FilterProxyModel:
+        return self._completion.model()
+
+    @property
+    def model(self) -> api.completion.BaseModel:
+        return self.proxy_model.sourceModel()
+
+    @property
+    def has_completions(self) -> bool:
+        return self.proxy_model.rowCount() > 0
 
     @utils.slot
     def _init_models(self):
         completionmodels.init()
 
-    @utils.slot
-    def _on_entered(self):
+    def initialize(self, text: str):
         """Initialize completion when command mode was entered."""
         # Set model according to text, defaults are not possible as
         # :command accepts arbitrary text as argument
-        self._update_proxy_model(self._cmd.text())
-        # Show if the model is not empty
-        self._maybe_show()
-        self._completion.raise_()
-
-    @utils.slot
-    def _on_left(self):
-        """Reset completion to empty model when leaving."""
-        self._proxy_model = api.completion.get_module("", api.modes.current())
+        self._update_proxy_model(text)
+        self._show_unless_empty()
 
     @utils.slot
     def _on_text_changed(self, text: str):
@@ -66,15 +66,14 @@ class Completer(QObject):
         self._completion.selectionModel().clear()
         # Update model
         self._update_proxy_model(text)
-        self._proxy_model.sourceModel().on_text_changed(
-            self._proxy_model.filtertext(text)
-        )
+        self.model.on_text_changed(text)
+        self._show_unless_empty()
 
     @utils.slot
     def _on_editing_finished(self):
         """Reset filter and hide completion widget."""
         self._completion.selectionModel().clear()
-        self._proxy_model.reset()
+        self.proxy_model.reset()
         self._completion.hide()
 
     def _update_proxy_model(self, text: str):
@@ -83,32 +82,23 @@ class Completer(QObject):
         Args:
             text: Text in the commandline which defines the model.
         """
-        proxy_model = api.completion.get_module(text, api.modes.COMMAND.last)
-        if proxy_model != self._proxy_model:
-            proxy_model.sourceModel().on_enter(proxy_model.filtertext(text))
-            self._proxy_model = proxy_model
-            self._completion.setModel(proxy_model)
+        model = api.completion.get_model(text, api.modes.COMMAND.last)
+        if model != self.model:
+            model.on_enter(text)
+            self.proxy_model.setSourceModel(model)
             self._completion.update_column_widths()
-        self._proxy_model.refilter(text)
-
-    def _maybe_show(self):
-        """Show completion widget if the model is not empty."""
-        if not isinstance(self._proxy_model.sourceModel(), completionmodels.Empty):
-            self._completion.show()
+        self.proxy_model.refilter(text)
 
     @utils.slot
-    def _on_completion(self, text: str):
-        """Set commandline text when completion was activated.
+    def _complete(self, text: str):
+        """Set commandline text including unmatched part (e.g. count) on completion."""
+        prefix, textpart = text[0], text[1:]
+        self._cmd.setText(prefix + self.proxy_model.unmatched + textpart)
 
-        Args:
-            text: Suggested text from completion.
-        """
-        # Get prefix and prepended digits
-        cmdtext = self._cmd.text()
-        prefix, cmdtext = cmdtext[0], cmdtext[1:]
-        digits = ""
-        while cmdtext and cmdtext[0].isdigit():
-            digits += cmdtext[0]
-            cmdtext = cmdtext[1:]
-        # Set text in commandline
-        self._cmd.setText(prefix + digits + text)
+    def _show_unless_empty(self):
+        """Show completion widget if there are completions, hide it otherwise."""
+        if not self.has_completions:
+            self._completion.hide()
+        elif not self._completion.isVisible():
+            self._completion.show()
+            self._completion.raise_()